src/textnode.py
from htmlnode import HtmlNode, LeafNode

class TextNode:
    def __init__(self, text, text_type, url = None): 
        self.text = text 
        self.text_type = text_type
        self.url = url

    def __eq__(self, other):
        return (
                self.text == other.text and
                self.text_type == other.text_type and
                self.url == other.url
        )

    def __repr__(self):
        return f"TextNode({self.text}, {self.text_type}, {self.url})"

InvalidTextTypeError = Exception

def text_node_to_html_node(text_node) -> HtmlNode:
    if text_node.text_type == "text":
        return LeafNode(None, text_node.text)
    if text_node.text_type == "bold":
        return LeafNode("b", text_node.text)
    if text_node.text_type == "italic":
        return LeafNode("i", text_node.text)
    if text_node.text_type == "code":
        return LeafNode("code", text_node.text)
    if text_node.text_type == "link":
        return LeafNode("a", text_node.text, {"href": text_node.url})
    if text_node.text_type == "image":
        return LeafNode("img", None, {"src": text_node.url, "alt": text_node.text})

    raise InvalidTextTypeError(text_node.text_type)

